refactor: reorganize auth module structure and improve integration handling
