+++ openspec/project.md
@@ -53,7 +53,7 @@ Nomad is a modern Online Travel Agency (OTA) built for budget-conscious, experie
 - **Server Actions:** `app/_actions` exposes Server Actions such as `auth.ts`, `flight-search-history.ts`, `orders.ts`, `payments.ts`, `passengers.ts`, and `dev-tools.ts`. They validate with Zod, call domain services, and return typed `ActionResult` objects.
 - **Shared Hooks & Components:** Domain-agnostic hooks (`app/_hooks`, e.g., `use-flight-search-state`, `use-otp-countdown`, `use-mobile`) keep logic reusable. Reusable UI elements live under `app/_components` (common layout pieces, domain-specific views, email templates, Fumadocs helpers, and the Shadcn-derived component library). `app/_components/common` includes `Header`, `Footer`, `DevUserSwitcher`, and `BreadcrumbNav`.
 - **Storybook:** `app/_stories` mirrors `app/_components` (auth, flights, forms, security, passengers, etc.) and ensures each shared component publishes at least one Storybook story.
-- **API Routes:** `app/api/auth/[...all]` delegates to Better Auth (`domains/auth/auth.ts`), `app/api/chat/route.ts` streams AI-powered answers, `app/api/cron/cancel-expiration` uses `src/security/cron` to authenticate and cancel expired orders, and `app/api/health/route.ts` returns a standardized `ApiResponse.success` payload validated against `src/types/api/health.ts`.
+- **API Routes:** `app/api/auth/[...all]` delegates to Better Auth (`domains/auth/infra/auth.plugin.ts`), `app/api/chat/route.ts` streams AI-powered answers, `app/api/cron/cancel-expiration` uses `src/security/cron` to authenticate and cancel expired orders, and `app/api/health/route.ts` returns a standardized `ApiResponse.success` payload validated against `src/types/api/health.ts`.
 - **Domain & Service Layers:** Business logic lives under `src/domains` (auth, booking, flights, payments, passengers, notification, user, dev-tools). Cross-domain orchestrations such as `src/services/payment-workflow.service.ts` handle balance payments, email notifications, and ledger updates.
 - **Integrations & Utilities:** `src/integrations` centralizes Aliyun SMS, Resend, Turnstile, and Fumadocs helpers (search, mermaid, LLM tooling). `src/lib` provides helpers for API responses, currency math, masking sensitive data, and logging.
 - **Database Schema:** `src/db/schema` defines tables for flights, airports, orders, passengers, and ancillary services; seeds are seeded via `drizzle-kit` and helper scripts.
